Building a DAO layer

+++ Chapter04/PostgreSQL/Maven/JpaGenericDao/src/main/java/com/classicmodels/jooq/repository/SaleRepository.java
@@ -11,5 +11,5 @@ import org.springframework.transaction.annotation.Transactional;
 public interface SaleRepository extends ClassicModelsRepository<SaleRecord, Sale, Long> {
 
     public List<Sale> findSaleByFiscalYear(int year);
-    public List<Sale> findSaleAscGtLimit(int limit);
+    public List<Sale> findSaleAscGtLimit(double limit);
 }
